+++ src/components/pages/Facilities/FacilitiesDetailPage/index.tsx
@@ -17,9 +17,9 @@ import Head from 'next/head';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { BsArrowLeftCircle } from 'react-icons/bs';
-import ReviewComponents from '@components/pages/reviews/ReviewComponents/reviewComponents';
-import CreateReviewModal from '@components/pages/reviews/createReviewModal';
-import UploadReviewImage from '@components/pages/reviews/uploadReviewImage';
+import ReviewComponents from '@components/models/reviews/ReviewComponents/reviewComponents';
+import CreateReviewModal from '@components/models/reviews/createReviewModal';
+import UploadReviewImage from '@components/models/reviews/uploadReviewImage';
 import UploadFacilityImage from '@components/pages/Facilities/FacilitiesDetailPage/UploadFacilityImage';
 import ReactStars from 'react-stars';
 import { supabase } from '@utils/supabaseClient';
